Major bugfix and new visualization features
Fixed a bug in the GUI where all tolerance units were read as m/z, meaning 10ppm would be treated as 10m/z tolerance.

Added ability to include axis tickmarks on publication-style figures to ease counting the number of pixels/line scans.

Added scaling to quantile for fractional images. In the GUI, scaling is to the percentile or using a threshold. This is made possible for all images and is made more clear.