Fixed TIC and interpolation bugs
There was a bug where the TIC for .raw files and Agilent .d files were potentially extracting the TIC from centroid data and the extracted intensities from profile data, leading to incorrect data. There was also a bug in interpolation that caused some images to be removed or duplicated. These have been fixed.