bugfix and added pixels_per_line parameter
Fixed a bug where normalize_img_sizes would be treated as True regardless of the input. This meant that output images would always have the same number of pixels.

Added the parameter "pixels_per_line". This allows users to state whether they want interpolation to use the minimum, maximum, or mean number of pixels across all line scans as the width of the image. Instead, an integer can be used to manually specify the desired width of the image in pixels.