Rewrote code to be contained in classes
Rewrote msigen.py, raw.py, mzml.py, and D.py to be classes, with msigen.py containing the base class that the others are subclasses of. This reduces the number of back and forth imports that are needed and makes it easier for all file types to share methods. It also reduces repeated functions and makes it easier to make future changes to the code.