Update Thermo Filter String Reading and related changes
Updated raw.py to read thermo scan filter strings more flexibly and reliably using regex. Including some support for SIM when there is only one SIM window used in the experiment. Removed some now pointless variables and methods.
get_PeakCountsPerFilter was renamed to get_CountsPerFilter. It was modified to better work with the changes to scan filter reading. Since the PeakCountsPerFilter variable was not used, it was removed everywhere.
consolidate_filter_list was updated to group together the scans where the precursors vary slightly in m/z, but are within the precursor m/z tolerance.