Added Parameters and changed backend
Fixed error that would occur when plotting images.

Added parameters and changed backend

Changed the matplotlib backend to agg to prevent crashing when generating large numbers of figures. This required the use of a new function to display the figures since agg is not interactive. pad_inches was added for changing the whitespace in the figure and dpi was added for changing the resolution of the figures. Also rewrote the plot_image function to reuse the same figure rather than recreating a figure for each image to increase speed and reduce memory use.